Практические работы/практика 6/requirements.txt
docx2pdf==0.1.8
pdf2docx==0.5.8
aspose-words==25.1.0
Pillow==11.1.0